Drop various/fixes/bad-font-gc patches
I have decided to drop the upstream debian
various/fixes/bad-font-gc set of revertion
patches.

It's possible that something about debian's
build configuration was responsible for the
problems that they had. I'm also unsure as
to whether we actually suffered from this
bug.

The reverted commits were large, and I'm not
too confident that this is the right way to
proceed, as upstream will over time be
marching in a different direction.