configure.sh: Divide SYS_CLANG > 0 section into two
The op_enable and gn_enable are moved further down
so that the clang checks and actual formating of
patches/flags and other things in rules/control
are separated.

Anything that can error out is done first, without
anything else getting in the way.

Since the whole section has been reformatted and
greatly simplified, the numerous comments are no
longer really needed.