configure.sh: Always prepend $(LLVM_DIR) onto AR/NM/RANLIB/CC/CXX
If we prepend $(LLVM_DIR) onto the toolchain
variable exports (AR/NM/RANLIB/CC/CXX) then we
can reduce the number of sed commands needed to
uncomment these exports.

Instead of matching 'llvm-' and 'clang', we just
need to  match '$(LLVM_DIR)/'. Two more commands
that prepended '$(LLVM_DIR)/' onto 'llvm-' and
'clang' are now unneeded.

We now need to make the value of the LLVM_DIR
export in d/rules take the value of LLVM_BASE_DIR
in the configuration script to accomodate setting
LLVM_DIR to /usr/local/bin.

We already use a sed command to substitute the
_LLVM_BASE_DIR marker in d/rules with with value
of $LLVM_BASE_DIR, so all we need to do is to
replace '/usr/lib/llvm-$(LLVM_VERSION)/bin'
with _LLVM_BASE_DIR/bin.

The export of LLVM_VERSION is now redundant, and
this will be removed in the next commit.