configure.sh: Set CLANG_VER to $CC_VER instead of $CR_VER
As noted in the previous commit, we can get rid of
the now redundant LLVM_VERSION in d/rules by reviving
CC_VER in d/configure.sh and setting CLANG_VER to the
value of this by default.

This also avoids any potential confusion between
LLVM_VERSION and LLVM_VER in both the rules file
and configuration script.

Also, let's stop using the '[ ]*' in sed expressions
which operate on d/control.in. It's technically
correct, and more fault tolerant, but far less
readable. As long as the dependencies are entered
correctly in d/control.in, there shouldn't be a
problem.