d/rules: Get rid of the tr command to get initial GN_FLAGS
I tried doing this a while back with a complicated
sed command which worked on the command line, but
failed in a real build situation.

I've got rid of the complication by using the '-z'
sed switch (a GNU extension), which allows the use
of '\n' newline notation.

The use of '\x' ASCII hex codes to prevent the
quotes or backslashes causing trouble seems to
allow the whole thing to actually work.