d/rules: Replace a cat command with a sed one
Although things seemed to work, it would be preferable
to set UB_LIST with a single line, instead of the list
of lines that the cat command was producing from
debian/misc_files/bundled.txt.

We can use part of the sed command from the previous
commit to return a single line list. A nice side
effect is that d/rules is now cat-less.