configure.sh: Allow using ESBuild javascript bundler
This is a new javascript bundler [1] that is
much faster than the default rollup.

It's still somewhat experimental, but should
speed up build times.

A patch is needed to remove an assert which
prevents official builds proceeding with the
devtools_fast_bundle build flag set tp true.

We introduce a new ESBUILD configuration
variable which is disabled by default (0).

Added :-
optional/enable-esbuild-for-official-builds.patch

[1] https://esbuild.github.io/