configure.sh: Make the POLLY/SYS_CLANG defaults consistent
Polly is disabled by default because it's not
compatible with the default of SYS_CLANG=0.

We had to specify POLLY=1 when SYS_CLANG > 0
with the optional polly patch being enabled by
default. So things weren't consistent.

Let's enable POLLY by default when SYS_CLANG > 0
while allowing it to be explicitly disabled under
that scenario. Make the patch disabled and enable
it when POLLY=1.

The POLLY=0 default and functional behaviour
should be unchanged.