configure.sh: Simplify sed expression behind CC_VER
Ironically, the commit where I tidied up similar
'[ ]*' notation is the one where I added this
line, after not fixing things up properly.

The new expression, beginning with either a hash
sign or a space, is a bit more flexible as it
should work regardless of whether bundled clang
is the default or not.