Fix build failure when SYS_ICU=0
This is caused by icudtl.dat being absent from
the build directory (out/Release). This file is
needed at least by top_domain_generator.

Add a commented line to d/rules.in which creates
a symlink in the build directory to the copy of
icudtl.dat in third_party/icu/common.

The configuration script has been modified to
uncomment this line when SYS_ICU=0.

The ln command is invoked with an '-f' switch
so that any pre-existing icudtl.dat in the
build directory gets clobbered by the symlink.