configure.sh: Enable {custom,host}_toolchain flags via gn_enable
Instead of using one GN_FLAGS assignation, we can
use two, which allows us to use gn_enable instead
of the RUL variable to unmask them in d/rules.

We also move these items further down d/rules.in
so that all the toolchain-related build flags
are in the same place.