Update upstream crash fix patches
All the patches below are from the various/upstream
folder.

0018-0a84a2b.diff was dropped because it's related
to testing, and I really want to focus on patching
the core browser functionality. Besides, there's
arguably too many patches already.

Patches 0016 and 0022 don't fix crashes, but are
needed by 0034 (with 0022 depending on 0016).

The number of patches has grown, but I expect that
several of the earlier patches will be obsolete
(and can be dropped) when 126 gets released. Most
patches are quite small and specific.

Dropped :-

0018-0a84a2b.diff

Renamed :-

0016-a45c3d5c.diff  --> 0017-a45c3d5c.diff
0017-7a735df58.diff --> 0018-7a735df58.diff

Added :-

0016-94cd472.diff
0019-3b82692.diff
0020-d6f2c59.diff
0021-e9b2942.diff
0022-9e4070dd.diff
0023-c999763.diff
0024-3e6cc50.diff
0025-2c69493.diff
0026-76f0ede2.diff
0027-17bdb17.diff
0028-38937ca.diff
0029-609320fa.diff
0030-5ddd7fa6.diff
0031-011f3e7.diff
0032-64b99f2.diff
0033-f61f79d4.diff
0034-3f8244f9.diff
0035-bb36794.diff
0036-a91265e3.diff
0037-1fbd955.diff
0038-e07a2713.diff
0039-4469b91.diff
0040-674acdce.diff