configure.sh: Never prune esbuild or pgo profile
Configuring with ESBUILD=1 after having configured
with ESBUILD=0 and started a build would result in
an error due to the esbuild binary not existing.

The correct binary is obtainable, but not easy to
find. Fix the issue by always exempting the binary
from the pruning process.

The same kind of problem is possible with the pgo
profile, so exempt this from pruning too.