d/rules: Fix ungoogled-chromium-l10n installation error
This bug was fixed a long while ago, but crept back in
with 676431b5d4885e07e8d0a5fcc13772b4e896637e, which
fixed a serious bug.

The latter bug was caused by an out-of-date en-US.pak
getting installed to the main package, causing a
glitched interface and stability issues.

We were checking for the existence of the en-US.pak
language file in the resources folder and not copying
it over (from the locales folder) if it existed. It
ended up never getting updated and eventually caused
the aforementioned problems.

Fixing this with an unconditional copy has caused
the old error (the en-US.pak file getting installed
by both the main and l10n packages) to re-appear.

Using a straight move at the end of the build instead
of a copy would solve the issue, but would mean having
to re-generate en-US.pak when doing a rebuild.

Instead we copy en-US.pak in the debian directory
after a build has completed. At the start of a build
we check to see if it exists in debian, and move it
to locales if it does.

Note that caching en-US.pak in the debian directory
ill avoid any problems with en-US.pak becoming out
of date.

Thanks to Tomtetom and ElEm4You at linuxmintusers.de
for reporting and identifying the issue.