Enable asynchronous leveldb protocol writes by default
An upstream patch provides the capability to enable
asynchronous writes for the leveldb protocol, and
another patch enables this by default.

The previous behaviour seems similar to mounting a
filesystem with synchronous writes enabled, which
is non-default on almost all journalled filesystems
due to the severe performance impact.

Added :-

various/leveldb-proto-add-feature-to-avoid-sync-writes.patch
various/leveldb-proto-enable-async-writes-by-default.patch