d/rules: Fix moving en-US.pak to debian directory
In a previous commit I used $(DEBIAN) instead
of $(DB_DIR) inside d/rules.in as the variable
for the debian directory.

The DEBIAN variable is from the configuration
script, so this screwed the commands in the
rules file which move en-US.pak around.

Fix this by renaming DB_DIR in d/rules.in to
DEBIAN.

Fixes: 32e5217664b180a40806b40f4a515dace113558a