shell launcher: Add ability to block runtime flags
We can add a flag file called 'blocked-flags', and
put any flags in there that we want to prohibit.

The format for this file is a space separated list
of flags, minus the double dash. If any of these
flags are found within CHROMIUM_FLAGS then we
remove them from this variable.