Workaround crash triggered by bookmarking a website
For some reason, the '--disable-sync' runtime flag
seems to cause a crash when bookmarking a site via
clicking on the star icon in the address bar.

The backtrace shows the crash happening in :-

base::ScopedObservation<syncer::SyncService, \
syncer::SyncServiceObserver>::Observe(syncer::SyncService*) ()

Although --disable-sync is not in our flag files,
add it to /etc/chromium.d/blocked-flags in order
to prevent crashes for those who set this flag.

Sync doesn't work in UC due to the feature being
tied to google services. So not overtly disabling
the feature via a flag shouldn't be a concern.