Fix an assert_no_deps configuration error
I'm not sure what changed between two recent
builds to trigger this error, but after much
tinkering I found that the fix is to remove
parts of an exisitng reversion patch.

The purpose of this reversion patch was to
avoid an increase in binary size, and the
bits removed only affect testing stuff which
won't have any affect on this.