shell launcher: Harden blocked flag processing
Blocked flags can be bypassed by specifying the
flag on the command line. This is arguably more
flexible as it allows the user to override the
block.

On the other hand, it's perhaps more useful from
a security perspective to not have this ability
to bypass.

Let's go with the latter by doing the blocked
flag processing directly after the positional
argument processing.

Anyone wanting to unblock a flag will now need
to have it removed from the blocked-flags file.