Disable --enable-webassembly-relaxed-simd by default
Enabling --enable-webassembly-relaxed-simd produces
a warning about disabling--expose_wasm due to
conflicting flags.

Using relaxed simd was supposed to produce more
speed at the expense of determinism. But having
a warning about, or actually disabling wasm is
not acceptable in a release.

Move this flag to the bottom of the webassembly
flag file and comment it out.