shell launcher: Allow matching a single isolated blocked flag
This is just to be ultra safe and correct.

Although we would almost always have more than
one flag in CHROMIUM_FLAGS, allow matching a
solitary flag when processessing blocked flags.

For the case statement, add '--$flag' to the
condition. Note that a solitary flag with a
single preceding or trailing space would be
picked up by the other two patterns.

For the sed command, we just need to add an
asterisk after the space in the first command
to match zero (or many) spaces.