shell launcher: Fix command line runtime flag blocking
The problem was that flags specified on the command line
(except for --enable-remote-extensions) weren't getting
included in the CHROMIUM_FLAGS variable. This was part
of the original design rather than a regression.

Fix by making anything beginning with a double dash and
a lower case letter to be appended to CHROMIUM_FLAGS.

Fixes: 09e3ebbc067e72dc90509fdcb88b0292e7397f4a