Silence some harmless error log messages
Silence 'Failed to decode OID' error logging by
downgrading it from LOG(ERROR) to LOG(INFO).
The latter should by default be hidden.

We rename silence-alsa-frames-log-msg.patch to
silence-some-error-log-messages.patch, and add
the change to this patch.

The alsa error logging is changed to downgrade
the log message in the same way, instead of
disabling it completely.