shell launcher: Process flag files before positional parameters
After some thought, it's probably better to allow
users to override anything in the flag files by
processing the positional parameters after the
flag files.

The last flag takes precedence, and there are pros
and cons as to what the order should be. However
the user should normally be able to override any
potential issues.

For example, we've just dealt with a crash caused
by the --disable-sync flag, so let's give users
in this situation the ability to avoid this crash
via setting --enable-sync.