configure.sh: Re-work sanitisation of CLANG_VER
CLANG_VER only needs to be sanitised when it is
specified by the user.

Let's do the sanitisation in the SYS_CLANG section
instead of having a separate sanitise_clang_ver
function, which isn't really needed.

Also, instead of trying to match a string with
anything containing a dot, be more specific by
requiring a two digit version number before the
dot.