configure.sh: Re-work sanitisation of op_disable/op_enable
We can use a case statement outside of the for
loop to check the whole op_disable/op_enable
string for 'optional/' prefixes, instead of
checking each individual patch within the
string.

We can then remove out each and every prefix
in one shot.

We don't need to call a function to do any of
this, and there's no real advantage to printing
any warnings either.