Move various/features to various/upstream folder
The various/features folder contains upstream patches
which are normally expected to be tied to some kind of
feature toggle in the configuration script, but are
not put in optional due to their transient nature.

Make the upstream nature of these patches clear by
locating their folder under various/upstream.