configure.sh: Fix SWITCH_BLOCKING default value
The default value (1) in configure-variables.txt
was what I had originally intended, but I somehow
missed changing the value in the configure.sh.

Default to SWITCH_BLOCKING=1 so that we at least
prevent '--disable-sync' from taking effect and
causing crashes.

Post-release, this shouldn't be a big problem
for 99 percent of people. In the meantime users
can edit /usr/bin/chromium to set SWITCH_BLOCKING
to their preferred value.