shell launcher: Allow specified flag files to be ignored
The new BLOCKED_FILES variable in the launcher can
be used to prevent specific flag files from being
sourced.

This could be useful as an alternative to commenting
out any uncommented flags in a flag file, as well as
for testing purposes.

The format for BLOCKED_FILES is a space-separated
list of file names, ie "file1 file2 ... fileN".

The whole for loop needs to be wrapped in an eval
command so that BLOCKED_FILES can be used in a
case statement condition.