Plumb in bindgen to the build system
Bindgen and the associated rust_bindgen_root build flag
are required for 127. The flag needs to be set to the
path that contains the bin directory containing the
bindgen binary.

There is no in-tree bindgen binary, and it's best for
builders to compile their own copy to ensure that it's
compatible with clang toolchain used to build UC.

To build we use the new build-bindgen.sh script which
has been added to to the (repo) build directory.

Refer to Bindgen.md for further details.