Fix lens build dependency issue when SYS_ICU=1
The problem is a dependency in a unit test that
that we don't build or use.

Since I'm running late in putting a release out
I have opted to just patch out the problematic
parts of a couple of BUILD.gn files, instead of
attempting a proper fix.

Added :-

optional/system/unstable/icu-lens-unittest-dep.patch