Clean up and correct bindgen handling and documentation
Having to integrate bindgen at short notice resulted
in things being rushed, and unfortunately quite a few
mistakes were made.

In the build-bindgen.sh script, the path to libclang.so
is third_party/rust-toolchain, and not the normal clang
library path (ie third_party/llvm-build/Release+Asserts).

Make SYS_CLANG optional for SYS_RUST=0 when, and error
out if (the former) is set to anything other than zero.

Set LIBCLANG_PATH to $CLANG_PATH/lib and use this to
set both LIBCLANG_PATH and LIBCLANG_STATIC_PATH in the
environment of the build command (LIBCLANG_STATIC_PATH
doesn't seem to be directly used in compilation).

In the documentation, SYS_CLANG=0 cannot be used with
SYS_RUST=1 when running build-bindgen.sh. SYS_CLANG=0
can only be used with SYS_RUST=0. List other valid
combinations too.

SYS_BINDGEN=0 is a leftover from a previous iteration
of configure.sh, and is now an invalid value.

SYS_BINDGEN=1 cannot be used with SYS_CLANG=0 or
SYS_CLANG=2, it can only be used with SYS_CLANG=1.

List other valid combinations of these two variables
for use in the configuration script.

The 'rules of thumb' for SYS_BINDGEN should have
referred to bindgen in /usr, instead of clang. But
let's omit this bit altogether, as things should be
clear enough from the earlier variable description.