widevine-cdm.sh: Fix execution of script from widevine-cdm folder
Firstly, rename the W_DIR variable to CUR_DIR. This is
a bit more descriptive (ie it stands for current dir).

The problem here is that CUR_DIR will evaluate to a blank
string when executing from the directory containing the
widevine-cdm.sh script.

Fix this by forcing a null CUR_DIR value to '.' ie the
shell notation for the current directory.