build-bindgen.sh: Fix regression and improve git handling
The '-c advice.detachedHead=false' argument to the git
clond of rust-bindgen got lost when rebasing, so add it
back in.

When updating an existing rust-bindgen checkout we were
changing into the rust-bindgen directory, then changing
back out, and then changing back in after exiting the
conditional.

Change into rust-bindgen under both branches of the
directory existence conditional, and stay there.

Also reverse the conditional so that we don't base
it on a negative condition.