build-bindgen: Fix ncurses downloading
The wget fallback capability was ported over from
the widevine download script, but unfortunately
the download caching wasn't wired up, which caused
downloading the ncurses zip file to fail.

Fix this by setting DL_CACHE to '.download_cache'
and ensuring that this folder exists.

We can now delete the DL_CACHE folder instead of
the zip file when running the script with the 'hc'
(or 'hardclean') positional parameter.