configure.sh: Exempt google support urls from domain substitution
These urls are of the form https://support.google.com/* and
are accessible via various places in the user interface.

They are set in chrome/common/url_constants.h, so add this
file to the other exemptions in configure.sh.

This should be safe, as we are only permitting specific
urls which a user has a choice whether to click or not.

The blocking of anything which is potentially averse to
privacy (eg tracking), won't be affected by this change.