configure.sh: Disable march and tune patches by setting MARCH=0
This is intended for those wanting to build on/for
ancient pre-2011 cpus.

This has actually uncovered a configuration bug, as
we need to set POLLY=0 when AVX=0, since POLLY depends
on AVX=1.

So move the bit that enables the polly patch to a point
after the cpu intruction patches.