d/rules: Stop automatically running ninja cleandead
Currently the ninja cleandead command is automatically
run at the end of a build.

Whilst this isn't necessarily a bad thing, it can slow
down build completion, and perhaps ought to be something
that users can choose to manually invoke.

Move the command so that it's run via directly invoking
debian/rules :-

./debian/rules cleandead