d/rules: Allow building with angle_enable_gl_desktop_frontend=true
Reading through gni/angle.gni seems to imply that the
'angle_enable_gl_desktop_frontend' build flag ought to
be enabled by default on linux.

Furthermore, the reason given for not enabling an
equivalent feature for windows is that it causes test
failures.

I'm not entirely sure what effect this has, but it
builds and runs fine. I'll allow users to decide
whether to enable it adding a GL_DESKTOP_FRONTEND
configuration variable.