d/rules: Build with libyuv_disable_rvv=true
In docs/environment_variables.md from chromium's libyuv
repo [1] it states that LIBYUV_DISABLE_RVV (the define
linked to the libyuv_disable_rvv build flag) is for
riscv cpus.

This seems to be an oversight from upstream, as other
similar flags are set according to cpu type. As we only
support x64 cpus, we can safely enable this flag and
perhaps save a few bytes of binary size.

[1] https://chromium.googlesource.com/libyuv/libyuv.git