Enable zero-copy VAAPI for vulkan and gl
For those who have working vaapi, this change should
increase decoding performance.

When configuring with VAAPI=1 (the default), two patches
are enabled, one for vulkan and one for GL. On vulkan
color depth is preserved (I'm not sure about GL).

For the GL version, the VaapiVideoDecodeLinuxZeroCopyGL
feature flag in /etc/chromium.d/hw-decoding-encoding
needs to be enabled. Note that this might be problematic
in certain setups/with particular drivers.

I have also updated /etc/chromium.d/hw-decoding-encoding
to try and keep up with whatever black magic is needed to
enable vaapi.

Added :-

optional/system/vaapi/zero-copy-for-VA-API-video-decoding-for-vulkan.patch
optional/system/vaapi/zero-copy-video-formats-with-GL-on-Linux.patch