configure.sh: Allow disabling dbus and notifictions system
The enable_system_notifications build flag was remove, but a
subsequent upstream commit [1] seemed to almost add back one
part of what had been removed, except it used the use_dbus
flag instead of the enable_system_notifications one.

From several unsuccessful attempts to build with use_dbus=false
in the previous release, it was apparent that notifications (of
all types) were dependent upon dbus.

The upshot of this is that I was able to partially revert the
commit which removed enable_system_notifications [2] but replace
each instance of enable_system_notifications with use_dbus.

In other words we can now employ the use_dbus flag as a proxy
for building without both dbus and notifications.

The stub-out-bluetooth-dbus patch is the same as the previous
atk-dbus one, minus the atk part.

I have added a DBUS configuration variable, and this silently
disables SYS_NOTIFICATIONS when set to zero.

NOTIFICATIONS is added too, and acts as an alias for DBUS since
use_dbus=false will disable both notifications and dbus itself.

The various/system/libdbus.patch when fixes building
with system dbus headers can now be made optional

Moved: various/system/libdbus.patch to optional/system/libdbus.patch

Added :-

optional/disable/dbus-and-notifications/fix-headless-build-issue-with-dbus-on-linux.patch
optional/disable/dbus-and-notifications/include-platform-util-linux-on-non-dbus.patch
optional/disable/dbus-and-notifications/notifications-depend-on-dbus.patch
optional/disable/dbus-and-notifications/stub-out-bluetooth-dbus.patch

[1] https://chromium-review.googlesource.com/c/chromium/src/+/5854776
[2] https://chromium-review.googlesource.com/c/chromium/src/+/5754151