configure.sh: Omit dawn_enable_vulkan=false if both WEBGPU=1 and VULKAN=1
I'm pretty sure the logic was wrong here because the
relevant line in d/rules.in is uncommented, yet we
have a gn_enable on the flag.

Also, the flag having both a webgpu (dawn) and vulkan
element implies that it shouldn't be enabled when
both WEBGPU or VULKAN are enaabled.

Re-jig this so that we use a gn_disable instead of a
gn_enable when both WEBGPU=1 and VULKAN=1.