configure.sh: Do not automatically enable CCACHE_BASEDIR
To enable it one needs to set the CCACHE_BASEDIR variable
to either 1 or a specific directory (absolute path).

Setting to 1 will set the variable to the root directory
of the source tree.

Not setting it, or setting it to zero, will do the same
as 1 except the export line will be commented out in
d/rules.

Note that we now have a separate line in d/rules to
export our /usr/lib/ccache prepended PATH.