Add ffmpeg hevc asm config
This patches an ffmpeg assembly config file [1] to enable
CONFIG_HEVCPARSE and CONFIG_HEVC_SEI. This was done after
referring to my own ffmpeg build to see which defines are
enabled for hevc.

I'm not sure whether this is strictly necessary, but it
probably won't do any harm since hevc asm files certainly
are included in the build.

Added :-

optional/ffmpeg-extra-codecs/hevc/add-hevc-ffmpeg-decoder-parser-for-linux-asm.patch

[1] third_party/ffmpeg/chromium/config/Chrome/linux/x64/config.asm