Enable building with ELF-CREL
This adds '-Wa,--crel,--allow-experimental-crel' to cflags
globally to enable the ELF CREL relocation format.

Enabling this reduces the size of ELF relocatable object
files (in out/Release/obj) by around 10%. While this has
no effect on binary size, it saves something in the order
of three quarters of a gigbyte in disk space.

This is pretty cutting edge, so is not enabled by default.
A clang toolchain containing a recent LLVM commit [1] is
needed to build with this option enabled.

A new ELF_CREL configuration variable is introduced to
toggle this feature.

More info at link references [2] and [3] below.

[1] https://github.com/llvm/llvm-project/commit/e82f0838ae88ad69515ebec234765e3e2607bebf
[2] https://discourse.llvm.org/t/rfc-crel-a-compact-relocation-format-for-elf/77600
[3] https://maskray.me/blog/2024-03-09-a-compact-relocation-format-for-elf