configure.sh: Fix/avoid ffmpeg extra codec clash issues
There's a clash in media/filters/ffmpeg_glue.cc in setting the
allowed_decoders variable when enabling extra audio codecs.

To solve this we work out which codecs are getting enabled via
the configuration script and use sed to write them into a single
patch [1]. This patch replaces the two hunks which alter the
allowed_decoders in the alac and fdk-aac patches.

Another potential clash is in the configuration headers within
third_party/ffmpeg [2], which is preventing the addition of AC3
decoding capability due to a potential context clash with ALAC.
Add an optional patch [3] to fix this, and have it disabled when
the internal FF_AUDIO variable is zero (ie when none of the extra
audio codecs are enabled).

As well as needing to adjust the context of an alac patch due
to [3], also bring the addition of the alac source files under
a separate 'if' heading, as is done with the equivalent fdk-aac
patch. This solves a potential context clash with hevc.

[1] optional/ffmpeg-extra-codecs/audio-codecs.patch
[2] chromium/config/Chrome/linux/x64/config_components.h
[3] optional/ffmpeg-extra-codecs/context-fixup.patch