Drop support for system ffmpeg
System ffmpeg is not recommended by upstream for security reasons
as it opens up a lot more potential attack surface than the intenal
chromium copy.

It's a bit of a kludge as there are often differences between the
internal and external implementations which can lead to problems
such as media not playing etc.

The patches are not really maintained or updated (by me), but have
been moved to the attic directory for anyone who finds them useful.