configure.sh: Make the Exec field in chromium.desktop run as a script
I have seen one report of the desktop file failing to
work (on Linux Mint). This was remedied by adding an
'sh -c' shell wrapper to contents of the Exec field.

Since our chromium executable is in fact a shell script
which launches the main executable, we ought to apply
this solution universally.

Added: various/fixes/desktop-exec-as-script.patch