configure.sh: Disable webassembly at runtime by default
I think it's probably best to by default lock down
components that have significant security concerns.

At the moment our defaults are to not build webgpu
(aka dawn) or pdf javascript functionality. Another
we ought to disable is webassembly.

This introduces a WEBASSEMBLY configuration variable
to toggle whether the '--js-flags=--noexpose_wasm'
runtime flag is enabled or not (default is zero).