configure.sh: Fix enabling system tiff when PDF_JS=1
The use_system_libtiff=true flag somehow went missing
from debian/rules, but is needed by pdf js/xfa forms.

The logic was also wrong in configure.sh, as the flag
needs to be enabled (not disabled) and should therefore
be acted on separately to the pdf flags.

This affects PDF_JS=1 only, which is not the default.