Add a couple of build fixes
The tint benchmarks patch fixes a build configuration
dependency error due to the tint_build_benchmarks build
flag being set to false.

The dbus secret portal patch fixes a symbol error for
those configuring with DBUS=0. The DbusSecretPortal
feature flag is disabled via this patch so that we are
not caught out when upstream enables it by default.

Added :-

various/disable/tint-benchmarks.patch
optional/disable/dbus-and-notifications/dbus-secret-portal-depends-on-dbus.patch