configure.sh: Stop POLLY_VEC from being undefined when POLLY=0
To avoid complexity, only exit with an error when
trying to enable POLLY with a bundled toolchain.

Other misconfigurations, such as setting AVX=0
and POLLY_VEC=1 will be silently corrected in
order to avoid complexity.

Enabling the polly vectorizer patch can be moved
up to be within the POLLY=1 conditional, to avoid
the exit on error.

Now the above has been clarified, we can simply
force POLLY_VEC=0 when AVX=0. POLLY_VEC is now
defined for both AVX=0 and AVX=1, and hence no
longer undefined further down in the script.