configure.sh: Fix/improve/clarify some clang issues
The bundled clang toolchain is (practically) by definition
compatible with PGO, and we need a way to check whether the
version of clang being used matches that of the bundled copy.

Previously CC_VER (the clang version used in debian/control)
was used for this purpose. But this variable is ultimately
meant to represent he latest clang/llvm version available
in debian unstable.

Since the latest clang in debian unstable will often lag the
latest used by chromium, we need a better solution.

Let's solve this by hardcoding the PGO compatible version as
LLVM_PGO_VER in the configuration script.

Also be clearer about where we switch from using CLANG_VER
to LLVM_VER. The former is set by the user, and is converted
to the latter after it has been sanitised and checked for PGO
compatibility.

Once LLVM_VER has been set universally, start using it in
the sed commands instead of continuing to use CLANG_VER.