d/rules: Update a couple of build flags
The glue_core_pools flag is now true by default, so
we can stop setting it in d/rules.

Setting enable_platform_hevc only makes sense when
setting it to false, as it's value depends on the
value of proprietary_codecs, which is always set
to true in our builds.

So reverse the logic and set enable_platform_hevc
to false whenever FF_HEVC=0.