d/rules: Do not set enable_swiftshader_vulkan=false by default
This slipped through the cracks and was left enabled in
d/rules even when swiftshader was disabled (SWIFTSHADER=0)
as the place where it got disable was under SWIFTSHADER=1.

Where the vulkan and webgpu swiftshader flags are set to
false (under SWIFTSHADER=1), I have changed things such
that each flag gets uncommented if the parent part (ie
vulkan or webgpu) is disabled, or the parent is enabled
together with disabling the respective swiftshader
component.

Note that the default for SWIFTSHADER_WEBPGU have been
changed from zero to one, even though WEBGPU is disabled
by default. The change just makes things easier to
implement.

We also stop force-enabling SWIFTSHADER_WEBGPU when
WEBGPU >= 1, since dawn_use_swiftshader depends on
swiftshader as per the webgpu gni file [1].

[1] third_party/dawn/scripts/dawn_features.gni