Slightly re-work clang version handling
If test mode is inactive (TEST=0 internally) then we
set LLVM_PGO_VER according the value of clang_version
in build/toolchain/toolchain.gni.

Move test mode detection further up the script in order
to accomodate this change.

We also don't need to set clang_version in d/rules when
LLVM_VER (or CLANG_VER from a user perspective) is set
to the same value as LLVM_PGO_VER. Print a little info
message when this is the case, since this will alert
myself and other users when clang_version gets bumped
upstream.