configure.sh: Re-work the enablement of google api keys
For those who don't use google services we can skip
the installation of the google-api-keys variable
export file [1].

Make the enablement of the actual exports depend on
setting GOOGLE_API_KEYS > 1. Also remove the reference
in google-api-keys to using sync, as this is crippled
in UC.

Non-builders will now have to uncomment the exports
manually if they want to use google services. Add a
note (in bold) to README.md to explain this.

[1] /etc/chromium.d/google-api-keys