configure.sh: Use ins_{en,dis}able instead of INS
I forgot that we have a simpler way to enable/disable
file installation, and what follows illustrates its
utility,

I screwed up the logic on disabling the installation of
google-api-keys. The INS expression was trying to enable
instead of disable installation, so use an ins_disable
in its place.

Also, the ins_disable under the LENS=0 conditional is
superfluous since LENS=0 is now the default.

Fixes: d696b838bb10d50e5b8a4d6d233796bfb7a2e352