Fix/re-work lto-CGO link optimisation
Looking to add '-Wl,--lto-CGO2' to where the '-O2'
optimisation switch is enabled, I discovered that the
'-Wl,--lto-CGO3' switch had been wrongly added to the
'-O2' config.

Re-work things so that the CGO level is inserted directly
after the respective optimisation switch, for both '-O2'
and '-O3'.

Since the '-O2' level is not being over-optimised (with
--lto-CGO3), this change should reduce binary size by
a small amount.

Relpace/rename the original patch [1] with [2], and tidy
up the context of optional/compiler-flags/polly.patch.

[1] various/compiler-flags/optimize_speed-plus-lto-CGO3.patch
[2] various/compiler-flags/optimize-lto-CGO.patch