shell launcher: Skip flag files via /etc/chromium.d/blocked-files
Instead of editing a list of flag files to ignore within
the shell launcher, use a blocklist file for this purpuse.

We use pipes ('|') instead of spaces as separators so that
the format is correct for the case statement where sourcing
flag files happens. Also add some basic sanitisation of the
BLOCKED_FILES variable.

By default the blocklist includes the blocked-files file
itself and the README file. We therefore always process
blocked-files, which allows the sourcing process to be
somewhat simplified.

Finally update /etc/chromium.d/README with details of the
new file blocking scheme.