Lower import hot multiplier from 15 to 14
A while back, a chromium developer estimated that the
ideal import limit was somewhere between 30 and 50.

At the moment our import limit of 450 (30x15) is
analagous to 45 at the default hot multiplier of
10 (45x10 = 450).

We'd probably be nearer the ideal if we went with
a multiplier of 14, which is analagous to an import
limit of 42 (30x14 = 42x10).

This is just a small step down and might save a
few bytes with negligable performance difference.