shell launcher: Fix strip_features function
This doesn't actually give an incorrect result in this
case, but we should be echoing into the strip_features
function rather than within it.

Also remove the double quotes from the CHROMIUM_FLAGS
variable, where we echo it into strip_features. These
aren't needed since the whole invocation is quoted.