shell launcher: Support --disable-web-security runtime flag
Some (misconfigured) websites fail to load correctly due
to Cross-Origin Resource Sharing (CORS) restrictions.
The --disable-web-security runtime flag can be used to
bypass these restrictions.

This flag requires --user-data-dir to be set, so alter
the shell launcher to set this when the former has been
specified (on the command line), but the latter hasn't.