Add a script to install node to third_party/node
This is for builders who have opted to use system
node and have applied DEPS-no-node.patch such that
no copy of node exists inside the third_party/node
directory.

The problem comes when sytem node doesn't work with
the bundled node modules, as is currently the case
with debian unstable's node.

After disabling optional/system/node.patch, running
the node-get.sh (from the 'build' directory) will
quickly restore a working copy of node, avoiding
the need to reset the build tree and run gclient
again.

The upstream bundled copy is listed as 20.11.0 in
DEPS, and the closest upstream version to this is
the 20.18.1 LTS upstream release (which appears
to work fine).