shell launcher: Stop exporting LD_LIBRARY_PATH
The whole launcher script was (afaik) inherited from
upstream debian, and it seems that the LD_LIBRARY_PATH
export section was originally intended to ensure that
libffmpeg.so was loaded.

We haven't built libffmpeg.so as separate library for a
long, long time, and since things seem to function fine
without exporting LD_LIBRARY_PATH, let's try living
without it.