Enable PartitionAllocFewerMemoryRegions feature by default
This is an upstream patch [1] which makes PartitionAlloc
use fewer (half or less) memory regions. Read through the
patch for technical details.

This requires the PartitionAllocFewerMemoryRegions feature
flag (in /etc/chromium.d/miscellaneous) to be enabled for
things to take effect (enabled by default).

Add the new PARTALLOC_MR configuration variable to allow
builders to disable the new flag.

[1] various/upstream/features/partitionalloc-add-feature-to-use-fewer-memory-regions.patch