Drop intel CET patches and rely on the buildflag
The patches were old and somewhat experimental, plus
there was (and still is) the enable_cet_shadow_stack
build flag, which has been enabled by default in the
period since the patches were introduced.

Change the INTEL_CET configuration variable default
from disabled to enabled, such that it acts as a
toggle for the aforementioned build flag (which
seems to mainly affect v8).