shell launcher: Fix quoted command line switches
Switches such as --profile-directory='Profile 2' (either
single or double-quoted) were being misprocessed due to
the switch value containing a space. The string afer the
space was (wrongly) getting interpreted as a url.

This issue was actually fixed by upstream debian, and
here we port their fix [1] to our launcher script.

[1] https://salsa.debian.org/chromium-team/chromium/-/commit/e8b6bbd59aadf18e41ee98841da7fc4783036957