Reland - configure.sh: Enable fast build restarts via FAST_RESTART
This relands 483d6f960b1d99e7a56069c9dad8657ac1562a24.

The difference between my earier reverted effort is that
we use the existence of out/Release/args.gn (instead of
the domain substitution cache file) to determine whether
a build is a first run or a restart.

We achieve this by deleting out/Release/args.gn during
configuration, and let the 'gn gen' command in d/rules
recreate it. On restarts this file will exist and the
'gn gen' command won't be run.

We also use a more finely tuned sed command in d/rules
during configuration to avoid altering the standalone
(manually invoked) version of the 'gn gen' command.

There appears to be some regeneration of the ninja build
configuaration as part of the build command, so it seems
likely that at least avoid some duplication is averted.